Ignore attempt to set new XHDOSLimits
According to a by-the-letter reading of the XHDI specification, XHDOSLimits
should only return EINVFN when the subfunction is not supported at all.
Otherwise it should always return the current limit. Thus, it is correct to
ignore attempts to set new limits.

Thank you to Uwe Seimet for pointing out this minutia (via issue #60).