Fix bug in ACSI I/O error handling
If an ACSI I/O completed with a non-zero status condition, the status
condition was just passed back to the caller of Rwabs(), resulting in
confusing error messages and stopping further I/O to the device.  We
now clear the status & ensure that a negative error code is returned.
Thanks to Vincent Riviere for reporting the bug (which was triggered
by a specific SD card with UltraSatan hardware).